Added, moved or updated some classes in the shared library.
