+++ Services/HardwareHero.Services.Shared/Middlewares/ExceptionHandlerMiddleware.cs
@@ -1,0 +1,40 @@
+﻿using HardwareHero.Services.Shared.Exceptions;
+using System.Net.Http;
+using System.Net;
+using Microsoft.AspNetCore.Http;
+
+namespace HardwareHero.Services.Shared.Middlewares
+{
+    public class ExceptionHandlerMiddleware
+    {
+        private readonly RequestDelegate _next;
+
+        public ExceptionHandlerMiddleware(RequestDelegate next)
+        {
+            _next = next;
+        }
+
+        public async Task InvokeAsync(HttpContext context)
+        {
+            try
+            {
+                await _next(context);
+            }
+            catch (DataValidationException ex)
+            {
+                context.Response.StatusCode = (int)HttpStatusCode.BadRequest;
+                await context.Response.WriteAsync(ex.Message);
+            }
+            catch (NotFoundException ex)
+            {
+                context.Response.StatusCode = (int)HttpStatusCode.NotFound;
+                await context.Response.WriteAsync(ex.Message);
+            }
+            //catch (AuthenticateException ex)
+            //{
+            //    context.Response.StatusCode = (int)HttpStatusCode.Unauthorized;
+            //    await context.Response.WriteAsync(ex.Message);
+            //}
+        }
+    }
+}
